Create Python Basics
a) Define a list with 20 elements and assign the elements with random integers
between (50, 100), sort the list and print the result. After sorting the list, calculate the 
sum of the elements at even locations (i.e., a[0] + a[2] + a[4] + a[6] +… + a[18])
(b) Calculate the mean, median, standard deviation of the list. Also, calculate the 2-based 
logarithm values of the list elements. Hint: use functions in modules statistics and math.